// lib/api/schemas/onboardingLevel.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */

export type OnboardingLevel = typeof OnboardingLevel[keyof typeof OnboardingLevel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OnboardingLevel = {
  BEGINNER: 'BEGINNER',
  INTERMEDIATE: 'INTERMEDIATE',
  ADVANCED: 'ADVANCED',
} as const;